Relax test runner's success criteria
Currently xsuccess is treated as a failure. But that will require the
tests on all platforms to have exactly the same result, either success
or failure. This requirement is looking infeasible for windows, where
many expected_fail tests succeed (probably succeeding in unintentional
ways) and many expected_succeed tests fail.